現在の進捗をpush
