+++ pages/index.vue
@@ -11,11 +11,8 @@ import { defineComponent, ref } from '@nuxtjs/composition-api'
 
 import IllustWindow from '@/components/templates/index/IllustWindow.vue'
 
-import InfoWindow from '@/components/templates/index/InfoWindow.vue'
-
 export default defineComponent({
   components: {
-    InfoWindow,
     IllustWindow,
   },
   layout: 'index',
